Distribute tools/tests.rs into inline test modules
Move tests from the monolithic src/tools/tests.rs into #[cfg(test)] mod tests
blocks colocated with the code they exercise:

- expand_path tests       -> tools/util.rs
- strip_frontmatter and
  resolve_agent_prompt    -> tools/agent.rs
- bash schema + async     -> tools/bash.rs
- read/write/edit/grep/
  find/ls schema + exec   -> respective tool files
- subagent oneshot/start  -> tools/subagent/{oneshot,start}.rs
- tool_registry_*         -> tools/registry.rs

The shared create_tool_context() helper now lives in a new
src/tools/test_helpers.rs module, gated by #[cfg(test)] and imported via
super::test_helpers::create_tool_context (or crate::tools::test_helpers
from subagent submodules).

No test logic changed; all 106 lib tests in the tools module pass.
