+++ src/tools/ls.rs
@@ -48,3 +48,43 @@ impl Tool for LsTool {
         }
     }
 }
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use super::super::test_helpers::create_tool_context;
+    use crate::tools::Tool;
+    use serde_json::json;
+
+    #[test]
+    fn test_ls_tool_schema() {
+        let tool = LsTool;
+        assert_eq!(tool.name(), "ls");
+        assert!(!tool.description().is_empty());
+
+        let params = tool.parameters();
+        assert_eq!(params["type"], "object");
+        assert!(params["properties"].is_object());
+        assert!(params["required"].is_array());
+    }
+
+    #[tokio::test]
+    async fn test_ls_tool_execution() {
+        let tool = LsTool;
+        let ctx = create_tool_context();
+
+        // Use a dedicated temp dir to avoid races with other tests
+        let dir = std::env::temp_dir().join("synaps_test_ls");
+        std::fs::create_dir_all(&dir).unwrap();
+        std::fs::write(dir.join("hello.txt"), "hi").unwrap();
+
+        let params = json!({
+            "path": dir.to_str().unwrap()
+        });
+
+        let result = tool.execute(params, ctx).await.unwrap();
+        assert!(result.contains("hello.txt"));
+
+        // Cleanup
+        std::fs::remove_dir_all(&dir).ok();
+    }
+}
